#27 refactoring structure, separate controller
