+++ src/main/java/com/dj/server/api/common/controller/MemberControllerAdvice.java
@@ -1,6 +1,6 @@
 package com.dj.server.api.common.controller;
 
-import com.dj.server.exception.member.MemberException;
+import com.dj.server.common.exception.member.MemberException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
