+++ src/main/java/com/dj/server/api/member/controller/MemberController.java
@@ -1,5 +1,10 @@
 package com.dj.server.api.member.controller;
 
+import com.dj.server.api.common.response.ResponseDTO;
+import com.dj.server.api.member.dto.request.KakaoProfile;
+import com.dj.server.api.member.dto.request.KakaoRequest;
+import com.dj.server.api.member.dto.request.KakaoToken;
+import com.dj.server.api.member.dto.response.ResponseTokenDTO;
 import com.dj.server.api.member.service.MemberService;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
@@ -10,6 +15,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 /**
@@ -29,8 +35,8 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 public class MemberController {
 
-    @Autowired
-    private MemberService memberService;
+    private final MemberService memberService;
+    private final KakaoRequest kakaoRequest;
 
     @ApiOperation(value = "test", notes = "테스트!")
     @ApiResponses({
@@ -40,6 +46,22 @@ public class MemberController {
     @GetMapping("/api/test")
     public String test() {
         return "123";
+    }
+
+
+    @ApiOperation(value = "signUp", notes = "로그인")
+    @ApiResponses({
+            @ApiResponse(code = 200, message = "OK"),
+            @ApiResponse(code = 404, message = "Not Found")
+    })
+    @GetMapping("/login/oauth2/kakao")
+    public ResponseDTO<ResponseTokenDTO> singUp(@RequestParam("code") String code, @RequestParam("redirect_url") String url) {
+        KakaoToken kakaoToken = kakaoRequest.getAccessToken(code, url);
+        KakaoProfile kakaoProfile = kakaoRequest.getKakaoProfile(kakaoToken);
+
+        // jwt
+        // 우리 서버가 생성한 jwt 토큰 두개를 같이 멤버에 넣어서 반환
+        return new ResponseDTO<>(memberService.getToken(kakaoProfile), "SUCCESS");
     }
 
 }
