+++ src/main/java/com/dj/server/api/member/service/MemberService.java
@@ -4,27 +4,11 @@ import com.dj.server.api.member.dto.request.KakaoProfile;
 import com.dj.server.api.member.dto.response.ResponseTokenDTO;
 import com.dj.server.api.member.entity.Member;
 import com.dj.server.api.member.entity.MemberRepository;
-import com.dj.server.api.member.entity.MemberRole;
-import com.dj.server.api.member.entity.SocialType;
 import com.dj.server.api.member.service.jwt.JwtUtil;
-import com.dj.server.exception.member.MemberCrudErrorCode;
-import com.dj.server.exception.member.MemberException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
-import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
-import org.springframework.security.oauth2.client.userinfo.OAuth2UserService;
-import org.springframework.security.oauth2.core.OAuth2AuthenticationException;
-import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import javax.servlet.http.HttpSession;
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * 유저에 대한 전반적인 비즈니스 로직을 담당합니다.
