+++ src/main/java/com/dj/server/common/aop/LoggingAop.java
@@ -1,4 +1,4 @@
-package com.dj.server.aop;
+package com.dj.server.common.aop;
 
 import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.ProceedingJoinPoint;
