+++ src/main/java/com/dj/server/common/interceptor/JwtConfig.java
@@ -1,4 +1,4 @@
-package com.dj.server.config.interceptor;
+package com.dj.server.common.interceptor;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
