+++ src/main/java/com/dj/server/test/controller/AopController.java
@@ -1,0 +1,30 @@
+package com.dj.server.test.controller;
+
+import com.dj.server.api.common.response.ResponseDTO;
+import com.dj.server.api.member.dto.response.MemberResponseDTO;
+import com.dj.server.test.dto.MemberTestRequestDTO;
+import lombok.ToString;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
+
+@ToString
+@RestController
+public class AopController {
+
+    /**
+     * aop 로그 확인을 위한 테스트 컨트롤러
+     *
+     * @param memberTestRequestDTO
+     * @return
+     */
+    @GetMapping("/test-aop/get")
+    @ResponseStatus(HttpStatus.OK)
+    public ResponseEntity<ResponseDTO<MemberResponseDTO>> testAop(@RequestBody MemberTestRequestDTO memberTestRequestDTO) {
+        return new ResponseEntity<>(new ResponseDTO<>(MemberResponseDTO.builder().memberSnsId(memberTestRequestDTO.getMemberSnsId())
+                .nickName(memberTestRequestDTO.getNickName()).build(), "SUCCESS"), HttpStatus.OK);
+    }
+}
