+++ src/main/java/com/dj/server/test/controller/CorsController.java
@@ -1,0 +1,32 @@
+package com.dj.server.test.controller;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.*;
+
+@RestController
+public class CorsController {
+
+    @PostMapping(value = {"/test-cors/post"})
+    @ResponseStatus(HttpStatus.OK)
+    public String testCorsPost() {
+        return "cors post passed!";
+    }
+
+    @GetMapping(value = {"/test-cors/get"})
+    @ResponseStatus(HttpStatus.OK)
+    public String testCorsGet() {
+        return "cors get passed!";
+    }
+
+    @PutMapping(value = {"/test-cors/put"})
+    @ResponseStatus(HttpStatus.OK)
+    public String testCorsPut() {
+        return "cors put passed!";
+    }
+
+    @DeleteMapping(value = {"/test-cors/delete"})
+    @ResponseStatus(HttpStatus.OK)
+    public String testCorsDelete() {
+        return "cors delete passed!";
+    }
+}
