+++ src/main/java/com/dj/server/test/dto/MemberTestRequestDTO.java
@@ -1,4 +1,4 @@
-package com.dj.server.api.member.dto.request;
+package com.dj.server.test.dto;
 
 import lombok.Builder;
 import lombok.Getter;
